Exception handling

+++ src/main/java/devsuperior/dscommerce/services/ProductService.java
@@ -4,16 +4,16 @@ package devsuperior.dscommerce.services;
 import devsuperior.dscommerce.dto.ProductDTO;
 import devsuperior.dscommerce.entities.Product;
 import devsuperior.dscommerce.repositories.ProductRepository;
+import devsuperior.dscommerce.services.exceptions.DatabaseException;
 import devsuperior.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.util.pattern.PathPatternRouteMatcher;
-
-import java.util.List;
-import java.util.Optional;
 
 @Service
 public class ProductService {
@@ -27,21 +27,8 @@ public class ProductService {
         return new ProductDTO(result);
     }
 
-
-
-    @Transactional(readOnly = true)
-    public ProductDTO findByIdErrado(Long id){
-        Optional<Product> result = repository.findById(id);
-        Product product = result.get();
-        ProductDTO dto = new ProductDTO(product);
-        return dto;
-    }
-
     @Transactional(readOnly = true)
     public Page<ProductDTO> findByAllErrado(Pageable pageable){
-        //List<Product> result = repository.findAll();
-        //return repository.findAll().stream().map(ProductDTO::new).toList();
-
         Page<Product> page = repository.findAll(pageable);
         return page.map(ProductDTO::new);
     }
@@ -53,16 +40,28 @@ public class ProductService {
         return new ProductDTO(product);
     }
 
-    @Transactional()
+    @Transactional
     public ProductDTO update(Long id, ProductDTO dto){
-        Product product = repository.getReferenceById(id);
-        copyProductDTOtoProduct(dto, product);
-        product = repository.save(product);
-        return new ProductDTO(product);
+       try {
+           Product product = repository.getReferenceById(id);
+           copyProductDTOtoProduct(dto, product);
+           product = repository.save(product);
+           return new ProductDTO(product);
+       }catch (EntityNotFoundException e){
+           throw new ResourceNotFoundException("Id not found");
+       }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException(("Id not exist"));
+        }
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("referential integrity failure");
+        }
         repository.deleteById(id);
     }
 
@@ -72,6 +71,4 @@ public class ProductService {
         product.setPrice(dto.price());
         product.setImgUrl(dto.imgUrl());
     }
-
-
 }
